+++ survey-service/src/main/java/com/ktpt/surmoon/service/survey/domain/model/item/question/select/SelectQuestionItemRepository.java
@@ -1,4 +1,4 @@
-package com.ktpt.surmoon.service.survey.domain.model.questionItem;
+package com.ktpt.surmoon.service.survey.domain.model.item.question.select;
 
 import org.springframework.data.jpa.repository.JpaRepository;
