+++ survey-service/src/main/java/com/ktpt/surmoon/service/survey/domain/model/item/question/text/TextQuestionItem.java
@@ -1,4 +1,4 @@
-package com.ktpt.surmoon.service.survey.domain.model.questionItem.text;
+package com.ktpt.surmoon.service.survey.domain.model.item.question.text;
 
 import java.time.LocalDateTime;
