+++ survey-service/src/main/java/com/ktpt/surmoon/service/survey/domain/model/item/question/text/TextQuestionItemType.java
@@ -1,4 +1,4 @@
-package com.ktpt.surmoon.service.survey.domain.model.questionItem.text;
+package com.ktpt.surmoon.service.survey.domain.model.item.question.text;
 
 public enum TextQuestionItemType {
     SHORT(50),
