+++ survey-service/src/test/java/com/ktpt/surmoon/service/survey/domain/model/common/SharedIdTest.java
@@ -1,16 +1,14 @@
 package com.ktpt.surmoon.service.survey.domain.model.common;
 
-import static org.assertj.core.api.Assertions.*;
-
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import com.ktpt.surmoon.service.survey.domain.model.questionItem.SelectQuestionItem;
-import com.ktpt.surmoon.service.survey.domain.model.questionItem.SelectQuestionItemRepository;
-import com.ktpt.surmoon.service.survey.domain.model.questionItem.text.TextQuestionItem;
-import com.ktpt.surmoon.service.survey.domain.model.questionItem.text.TextQuestionItemRepository;
-import com.ktpt.surmoon.service.survey.domain.model.questionItem.text.TextQuestionItemType;
+import com.ktpt.surmoon.service.survey.domain.model.item.question.select.SelectQuestionItem;
+import com.ktpt.surmoon.service.survey.domain.model.item.question.select.SelectQuestionItemRepository;
+import com.ktpt.surmoon.service.survey.domain.model.item.question.text.TextQuestionItem;
+import com.ktpt.surmoon.service.survey.domain.model.item.question.text.TextQuestionItemRepository;
+import com.ktpt.surmoon.service.survey.domain.model.item.question.text.TextQuestionItemType;
 
 @DataJpaTest
 public class SharedIdTest {
